cuda: numerical accuracy round 1 — router probs, sinks, HC expand
Three structural mismatches uncovered by ./ds4 --metal-graph-test that
the CUDA backend now matches CPU reference for:

1. Router probability formula.  The DSV4 router uses
       probs[e] = sqrt(softplus(logits[e]))
   followed by top-K by (probs + bias) and weights[k] =
   probs[sel[k]] / max(sum, eps) * 1.5.  The CUDA implementation was
   doing a plain softmax over the logits.  cuda/router.cuh now mirrors
   metal/unary.metal::kernel_dsv4_softplus_sqrt and
   metal/dsv4_misc.metal::kernel_dsv4_router_weights_one.

2. Attention sinks tensor type.  The model stores per-head sinks as
   F32 (Metal does `(const float *)sinks`) but the CUDA backend was
   reading them as __half, which both produced wrong values and let
   the kernel walk 2x past the registered range.  Fixed across all six
   attention variants and the size check in ds4_cuda.cu.

3. HC expand combination matrix layout.  The split tensor per token is
       [pre[n_hc], post[n_hc], comb[n_hc x n_hc]]
   with comb's fastest-changing dim being dst_hc (Metal nb_comb0 =
   sizeof(float), nb_comb1 = n_hc * sizeof(float)).  The CUDA kernel
   was reading post at the pre-weight offset (off by n_hc floats) and
   ignoring comb entirely, replacing the full HC mix with a flat
   residual add.  ds4_metal_hc_expand_split_tensor and
   _add_split_tensor both rewritten to do the proper Metal math.

Effect, measured against --metal-graph-test layer 0 diffs:
    after_attn_hc    4.69    -> 0.003   (-1500x)
    ffn_cur          9.83    -> 0.003
    logits          40.64    -> 4.52

The remaining error is dominated by routed_moe (diff 18.5).  Router
probs/weights now match the math in ds4.c's CPU reference, but
upstream numerical drift (q_rope 0.032, kv_rope 0.053) lets the top-K
ranking flip across experts.  Next round will tighten RoPE and attention.

+++ cuda/attention.cuh
@@ -50,7 +50,7 @@ __device__ __forceinline__ float ds4_cuda_attn_block_sum(float val, float *smem_
  *   window:   sliding-window length (token attends to [t-window+1 .. t]) */
 template <int BLOCK>
 __global__ void ds4_cuda_kernel_attn_prefill_raw_f32(
-        float *heads, const float *q, const float *raw_kv, const __half *sinks,
+        float *heads, const float *q, const float *raw_kv, const float *sinks,
         uint32_t n_tok, uint32_t window, uint32_t n_head, uint32_t head_dim) {
     const uint32_t t = blockIdx.x;
     const uint32_t h = blockIdx.y;
@@ -70,7 +70,7 @@ __global__ void ds4_cuda_kernel_attn_prefill_raw_f32(
     float *warp_buf = smem;            /* 32 entries */
     float *score_bcast = smem + 32;    /* 1 entry, reused */
 
-    const float sink = sinks ? __half2float(sinks[h]) : -FLT_MAX;
+    const float sink = sinks ? sinks[h] : -FLT_MAX;
 
     /* Pass 1: max logit over the window plus the attention sink. */
     float max_logit = sink;
@@ -112,7 +112,7 @@ __global__ void ds4_cuda_kernel_attn_prefill_raw_f32(
  * absolute position. */
 template <int BLOCK>
 __global__ void ds4_cuda_kernel_attn_decode_raw_batch_f32(
-        float *heads, const float *q, const float *raw_kv, const __half *sinks,
+        float *heads, const float *q, const float *raw_kv, const float *sinks,
         uint32_t n_tok, uint32_t pos0,
         uint32_t n_raw, uint32_t raw_cap, uint32_t raw_start,
         uint32_t window, uint32_t n_head, uint32_t head_dim) {
@@ -134,7 +134,7 @@ __global__ void ds4_cuda_kernel_attn_decode_raw_batch_f32(
     extern __shared__ float smem[];
     float *warp_buf = smem;
     float *score_bcast = smem + 32;
-    const float sink = sinks ? __half2float(sinks[h]) : -FLT_MAX;
+    const float sink = sinks ? sinks[h] : -FLT_MAX;
 
     float max_logit = sink;
     for (uint32_t j = 0; j < take; j++) {
@@ -189,7 +189,7 @@ __global__ void ds4_cuda_kernel_attn_decode_raw_batch_f32(
 template <int BLOCK>
 __global__ void ds4_cuda_kernel_attn_prefill_static_mixed_f32(
         float *heads, const float *q, const float *raw_kv, const float *comp_kv,
-        const __half *sinks,
+        const float *sinks,
         uint32_t n_tok, uint32_t n_comp, uint32_t window, uint32_t ratio,
         uint32_t n_head, uint32_t head_dim) {
     const uint32_t t = blockIdx.x;
@@ -214,7 +214,7 @@ __global__ void ds4_cuda_kernel_attn_prefill_static_mixed_f32(
     extern __shared__ float smem[];
     float *warp_buf = smem;
     float *score_bcast = smem + 32;
-    const float sink = sinks ? __half2float(sinks[h]) : -FLT_MAX;
+    const float sink = sinks ? sinks[h] : -FLT_MAX;
 
     /* Pass 1: max. */
     float max_l = sink;
@@ -274,7 +274,7 @@ __global__ void ds4_cuda_kernel_attn_prefill_static_mixed_f32(
 template <int BLOCK>
 __global__ void ds4_cuda_kernel_attn_prefill_masked_mixed_f32(
         float *heads, const float *q, const float *raw_kv, const float *comp_kv,
-        const float *comp_mask, const __half *sinks,
+        const float *comp_mask, const float *sinks,
         uint32_t n_tok, uint32_t n_comp, uint32_t window, uint32_t ratio,
         uint32_t n_head, uint32_t head_dim) {
     (void)ratio;
@@ -293,7 +293,7 @@ __global__ void ds4_cuda_kernel_attn_prefill_masked_mixed_f32(
     extern __shared__ float smem[];
     float *warp_buf = smem;
     float *score_bcast = smem + 32;
-    const float sink = sinks ? __half2float(sinks[h]) : -FLT_MAX;
+    const float sink = sinks ? sinks[h] : -FLT_MAX;
 
     float max_l = sink;
     for (int32_t j = jraw_lo; j <= jraw_hi; j++) {
@@ -353,7 +353,7 @@ __global__ void ds4_cuda_kernel_attn_prefill_masked_mixed_f32(
 template <int BLOCK>
 __global__ void ds4_cuda_kernel_attn_decode_mixed_batch_f32(
         float *heads, const float *q, const float *raw_kv, const float *comp_kv,
-        const float *comp_mask, uint32_t use_comp_mask, const __half *sinks,
+        const float *comp_mask, uint32_t use_comp_mask, const float *sinks,
         uint32_t n_tok, uint32_t pos0, uint32_t n_raw, uint32_t raw_cap, uint32_t raw_start,
         uint32_t n_comp, uint32_t window, uint32_t ratio, uint32_t n_head, uint32_t head_dim) {
     (void)ratio;
@@ -373,7 +373,7 @@ __global__ void ds4_cuda_kernel_attn_decode_mixed_batch_f32(
     extern __shared__ float smem[];
     float *warp_buf = smem;
     float *score_bcast = smem + 32;
-    const float sink = sinks ? __half2float(sinks[h]) : -FLT_MAX;
+    const float sink = sinks ? sinks[h] : -FLT_MAX;
 
     float max_l = sink;
     for (uint32_t j = 0; j < take; j++) {
@@ -435,7 +435,7 @@ __global__ void ds4_cuda_kernel_attn_decode_mixed_batch_f32(
 template <int BLOCK>
 __global__ void ds4_cuda_kernel_attn_indexed_mixed_batch_f32(
         float *heads, const float *q, const float *raw_kv, const float *comp_kv,
-        const int32_t *topk, const __half *sinks,
+        const int32_t *topk, const float *sinks,
         uint32_t n_tok, uint32_t pos0, uint32_t n_raw, uint32_t raw_cap, uint32_t raw_start,
         uint32_t n_comp, uint32_t top_k, uint32_t window, uint32_t ratio,
         uint32_t n_head, uint32_t head_dim) {
@@ -456,7 +456,7 @@ __global__ void ds4_cuda_kernel_attn_indexed_mixed_batch_f32(
     extern __shared__ float smem[];
     float *warp_buf = smem;
     float *score_bcast = smem + 32;
-    const float sink = sinks ? __half2float(sinks[h]) : -FLT_MAX;
+    const float sink = sinks ? sinks[h] : -FLT_MAX;
 
     float max_l = sink;
     for (uint32_t j = 0; j < take; j++) {
